Split app.js into multiple files
