+++ frontend/src/App.js
@@ -1,408 +1,11 @@
 import React, { useState, useEffect, use } from "react";
 import "./style.css";
 import { ManageUsersModal } from "./ManageUsersModal";
-
-const ChemicalModal = ({ chemical, show, handleClose }) => {
-  if (!chemical) return null; // Don't render if no chemical is selected
-
-  return (
-    <div
-      className={`modal fade ${show ? "show d-block" : "d-none"}`}
-      tabIndex="-1"
-    >
-      <div className="modal-dialog modal-xl modal-dialog-centered modal-dialog-scrollable">
-        <div className="modal-content">
-          <div className="modal-header">
-            <h1 className="modal-title fs-5">{chemical.name}</h1>
-            <button
-              type="button"
-              className="btn-close"
-              onClick={handleClose}
-            ></button>
-          </div>
-          <div className="modal-body">
-            <label className="form-label">Chemical Abbreviation</label>
-            <input
-              type="text"
-              className="form-control"
-              value={chemical.symbol}
-              readOnly
-            />
-            <label className="form-label">Storage Class</label>
-            <input
-              type="text"
-              className="form-control"
-              placeholder="Corr White"
-              readOnly
-            />
-            <label className="form-label">MSDS</label>
-            <input
-              type="text"
-              className="form-control"
-              placeholder="MSDS"
-              readOnly
-            />
-            <label className="form-label">Minimum Needed</label>
-            <input
-              type="text"
-              className="form-control"
-              placeholder="3"
-              readOnly
-            />
-            <table className="table">
-              <thead>
-                <tr>
-                  <th scope="col">Sticker #</th>
-                  <th scope="col">Product #</th>
-                  <th scope="col">Location</th>
-                  <th scope="col">Sub-Location</th>
-                  <th scope="col">Manufacturer</th>
-                  <th scope="col">Quantity</th>
-                </tr>
-              </thead>
-              <tbody>
-                {chemical.inventory?.map((item, index) => (
-                  <tr key={index}>
-                    <th scope="row">{item.sticker}</th>
-                    <td>{item.product}</td>
-                    <td>{item.location}</td>
-                    <td>{item.subLocation}</td>
-                    <td>{item.manufacturer}</td>
-                    <td>{item.quantity}</td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-          <div className="modal-footer">
-            <button
-              type="button"
-              className="btn btn-secondary"
-              onClick={handleClose}
-            >
-              Close
-            </button>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-};
-const Navbar = ({handleShowAddChemicalModal}) => {
-  const [user, setUser] = useState({});
-
-  useEffect(() => {
-    fetch("/api/user")
-      .then((response) => response.json())
-      .then((data) => setUser(data))
-      .catch((error) => console.error(error));
-  }, []);
-  return (
-    <nav className="navbar navbar-expand-lg bg-light">
-      <div className="container-fluid">
-        <a className="navbar-brand me-auto" href="/">
-          Chemical Inventory
-        </a>
-        <div className="collapse navbar-collapse" id="navbarSupportedContent">
-                    <ul className="navbar-nav me-auto mb-2 mb-lg-0">
-                        <li className="nav-item">
-                            <a
-                                className="nav-link active"
-                                href="#"
-                                onClick={(e) => {
-                                    e.preventDefault();
-                                    handleShowAddChemicalModal();
-                                }}
-                            >
-                                Add Chemical
-                            </a>
-                        </li>
-                    </ul>
-          <ul className="navbar-nav mb-2 mb-lg-0">
-            <li className="nav-item dropdown">
-              <a
-                className="nav-link dropdown-toggle"
-                href="#"
-                id="navbarDropdown"
-                role="button"
-                data-bs-toggle="dropdown"
-                aria-expanded="false"
-              >
-                Hi {user.name}
-              </a>
-              <ul
-                className="dropdown-menu dropdown-menu-end"
-                aria-labelledby="navbarDropdown"
-              >
-                <li className="dropdown-item">You have {user.access} access</li>
-                {user.access === "admin" ? (
-                  <li>
-                    <a
-                      className="dropdown-item"
-                      data-bs-toggle="modal"
-                      data-bs-target="#manageUsersModal"
-                    >
-                      Manage access
-                    </a>
-                  </li>
-                ) : null}
-                <li>
-                  <hr className="dropdown-divider" />
-                </li>
-                <li>
-                  <a className="dropdown-item" href="#">
-                    Logout
-                  </a>
-                </li>
-              </ul>
-            </li>
-          </ul>
-        </div>
-      </div>
-    </nav>
-  );
-};
-
-const AddChemicalModal = ({show, handleClose}) => {
-    return (
-        <div className={`modal fade ${show ? "show d-block" : "d-none"}`} tabIndex="-1">
-            <div className="modal-dialog modal-xl modal-dialog-centered modal-dialog-scrollable">
-                <div className="modal-content">
-                    <div className="modal-header">
-                        <h1 className="modal-title fs-5" id="addChemicalLabel">Acetic Acid</h1>
-                        <button type="button" className="btn-close" data-bs-dismiss="modal" aria-label="Close" onClick={handleClose}></button>
-                    </div>
-                    <div className="modal-body">
-                        {/* Identification Section */}
-                        <div className="grouped-section">
-                            <label className="form-label">Sticker Number</label>
-                            <input type="text" className="form-control" placeholder="3008" readOnly/>
-                            <label className="form-label">Chemical Name</label>
-                            <input type="text" className="form-control" placeholder="Acetic Acid" readOnly/>
-                            <label className="form-label">Chemical Formula/Common Name</label>
-                            <input type="text" className="form-control" placeholder="" readOnly/>
-                            <label className="form-label">Storage Class</label>
-                            <input type="text" className="form-control" placeholder="Corr White" readOnly/>
-                            <button type="button" className="btn btn-secondary">Select Chemical</button>
-                            <button type="button" className="btn btn-secondary">Fix Chemical Typo</button>
-                        </div>
-
-                        {/* Quantity Section */}
-                        <div className="grouped-section">
-                            <label className="form-label">Quantity</label>
-                            <input type="text" className="form-control" placeholder="1"/>
-                            <label className="form-label">Unit</label>
-                            <select className="form-select">
-                                <option selected>Bottle</option>
-                                <option value="Liters">Liters</option>
-                                <option value="Milliliters">Milliliters</option>
-                                <option value="Kilograms">Kilograms</option>
-                            </select>
-                        </div>
-
-                        {/* Update Info Section */}
-                        <div className="grouped-section">
-                            <label className="form-label">Last Updated</label>
-                            <input type="text" className="form-control" placeholder="2024-09-19" readOnly/>
-                            <label className="form-label">Updated By</label>
-                            <input type="text" className="form-control" placeholder="emowat" readOnly/>
-                        </div>
-
-                        {/* Location Section */}
-                        <div className="grouped-section">
-                            <label className="form-label">Location</label>
-                            <select className="form-select">
-                                <option selected>FC 212</option>
-                                <option value="FC 101">FC 101</option>
-                                <option value="Lab 3">Lab 3</option>
-                            </select>
-                            <label className="form-label">Sub-Location</label>
-                            <select className="form-select">
-                                <option selected>Hood G</option>
-                                <option value="Shelf A">Shelf A</option>
-                                <option value="Cabinet B">Cabinet B</option>
-                            </select>
-                        </div>
-
-                        {/* Manufacturer Section */}
-                        <div className="grouped-section">
-                            <label className="form-label">Manufacturer Name</label>
-                            <select className="form-select">
-                                <option selected>TCI</option>
-                                <option value="Sigma-Aldrich">Sigma-Aldrich</option>
-                                <option value="Fisher Scientific">Fisher Scientific</option>
-                            </select>
-                            <label className="form-label">Product Number</label>
-                            <input type="text" className="form-control" placeholder="N0155"/>
-                            <label className="form-label">Material Safety Data Sheet</label>
-                            <input type="text" className="form-control" placeholder="MSDS Link" readOnly/>
-                            <button type="button" className="btn btn-secondary">Add Manufacturer</button>
-                        </div>
-                    </div>
-                    <div className="modal-footer">
-                        <button type="button" className="btn btn-secondary" data-bs-dismiss="modal"
-                                aria-label="Close" onClick={handleClose}>Cancel
-                        </button>
-                        <button type="button" className="btn btn-primary">Save Chemical</button>
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-};
-
-const Sidebar = ({
-  chemicals,
-  rooms,
-  manufacturers,
-  query,
-  setQuery,
-  handleSearch,
-}) => {
-  const [selectedChemicals, setSelectedChemicals] = useState([]);
-  const [selectedRoom, setSelectedRoom] = useState("");
-  const [selectedManufacturers, setSelectedManufacturers] = useState([]);
-
-  const toggleChemical = (chem) => {
-    setSelectedChemicals((prev) =>
-      prev.includes(chem) ? prev.filter((c) => c !== chem) : [...prev, chem]
-    );
-  };
-
-  const toggleManufacturer = (man) => {
-    setSelectedManufacturers((prev) =>
-      prev.includes(man) ? prev.filter((m) => m !== man) : [...prev, man]
-    );
-  };
-
-  return (
-    <div className="tw-w-1/4 tw-bg-white tw-p-4 tw-rounded-md tw-shadow-md">
-      <div className="tw-flex tw-items-center tw-border tw-p-2 tw-rounded-md">
-        <form
-          className="tw-w-full tw-flex"
-          onSubmit={(e) => {
-            e.preventDefault();
-            handleSearch(query);
-          }}
-        >
-          <input
-            name="query"
-            type="text"
-            placeholder="Search..."
-            className="tw-ml-2 tw-w-full tw-outline-none"
-            value={query}
-            onChange={(e) => setQuery(e.target.value)}
-          />
-          <button className="tw-flex tw-items-center tw-justify-center tw-w-8 tw-h-8 tw-bg-transparent tw-border-none hover:tw-opacity-70">
-            <span className="material-icons">search</span>
-          </button>
-        </form>
-      </div>
-      <div className="tw-mt-4">
-        <div className="tw-font-semibold">Popular Chemicals</div>
-        <div className="tw-mt-2 tw-space-y-1">
-          {chemicals.map((chem, index) => (
-            <label key={index} className="tw-flex tw-items-center">
-              <input
-                type="checkbox"
-                className="tw-mr-2"
-                checked={selectedChemicals.includes(chem)}
-                onChange={() => toggleChemical(chem)}
-              />
-              {chem}
-            </label>
-          ))}
-        </div>
-      </div>
-      <div className="tw-mt-4">
-        <div className="tw-font-semibold">Room Location</div>
-        <div className="tw-mt-2 tw-space-y-1">
-          {rooms.map((room, index) => (
-            <label key={index} className="tw-flex tw-items-center">
-              <input
-                type="radio"
-                name="room"
-                className="tw-mr-2"
-                checked={selectedRoom === room}
-                onChange={() => setSelectedRoom(room)}
-              />
-              {room}
-            </label>
-          ))}
-        </div>
-      </div>
-      <div className="tw-mt-4">
-        <div className="tw-font-semibold">Manufacturers</div>
-        <div className="tw-mt-2 tw-space-y-1">
-          {manufacturers.map((man, index) => (
-            <label key={index} className="tw-flex tw-items-center">
-              <input
-                type="checkbox"
-                className="tw-mr-2"
-                checked={selectedManufacturers.includes(man)}
-                onChange={() => toggleManufacturer(man)}
-              />
-              {man}
-            </label>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-const MainContent = ({
-  chemicalsData,
-  loading,
-  query,
-  handleSearch,
-  handleShowModal,
-}) => (
-  <div className="tw-w-3/4 tw-bg-white tw-ml-4 tw-p-4 tw-rounded-md tw-shadow-md">
-    <div className="tw-grid tw-grid-cols-3 tw-border-b tw-p-2 tw-font-semibold">
-      <div>Quantity</div>
-      <div>Chemical</div>
-      <div>Chemical Symbol</div>
-    </div>
-    <div className="tw-divide-y">
-      {loading && chemicalsData.length === 0 ? (
-        <p>Loading...</p>
-      ) : (
-        chemicalsData.map((chem, index) => (
-          <div key={index} className="tw-grid tw-grid-cols-3 tw-p-2">
-            {/*Columns on the main page*/}
-            <div>{chem.quantity}</div>
-            <div>
-              <a
-                href="#"
-                className="text-primary text-decoration-none"
-                onClick={(e) => {
-                  e.preventDefault();
-                  handleShowModal(chem);
-                }}
-              >
-                {chem.chemical_name}
-              </a>
-            </div>
-            <div>{chem.formula}</div>
-          </div>
-        ))
-      )}
-    </div>
-    <div className="d-flex justify-content-center">
-      {query !== "" && (
-        <button
-          className="btn btn-outline-success mt-3 mx-auto"
-          type="submit"
-          onClick={() => handleSearch(query, true)}
-          disabled={loading}
-        >
-          {loading ? "Searching..." : "Expand Search"}
-        </button>
-      )}
-    </div>
-  </div>
-);
+import { Sidebar } from "./Sidebar";
+import { MainContent } from "./MainContent";
+import { Navbar } from "./Navbar";
+import { AddChemicalModal } from "./AddChemicalModal";
+import { ChemicalModal } from "./ChemicalModal";
 
 const App = () => {
   const [query, setQuery] = useState("");
@@ -413,7 +16,7 @@ const App = () => {
   const [selectedChemical, setSelectedChemical] = useState(null);
   const [showModal, setShowModal] = useState(false);
 
-    const [showAddChemicalModal, setShowAddChemicalModal] = useState(false);
+  const [showAddChemicalModal, setShowAddChemicalModal] = useState(false);
 
   const handleShowModal = (chem) => {
     setSelectedChemical(chem);
@@ -424,13 +27,13 @@ const App = () => {
     setShowModal(false);
   };
 
-    const handleShowAddChemicalModal = () => {
-        setShowAddChemicalModal(true);
-    };
+  const handleShowAddChemicalModal = () => {
+    setShowAddChemicalModal(true);
+  };
 
-    const handleCloseAddChemicalModal = () => {
-        setShowAddChemicalModal(false);
-    };
+  const handleCloseAddChemicalModal = () => {
+    setShowAddChemicalModal(false);
+  };
 
   function handleSearch(query, synonyms = false) {
     if (query === "") {
@@ -476,7 +79,6 @@ const App = () => {
   useEffect(() => {
     getQuantity(); // Fetch chemicals on component mount
   }, []);
-
 
   useEffect(() => {
     const handler = setTimeout(() => {
@@ -548,7 +150,7 @@ const App = () => {
 
   return (
     <div className="tw-bg-gray-100 pb-3">
-            <Navbar handleShowAddChemicalModal={handleShowAddChemicalModal}/>
+      <Navbar handleShowAddChemicalModal={handleShowAddChemicalModal} />
       <div className="tw-flex tw-mt-4">
         <Sidebar
           chemicals={chemicals}
@@ -567,11 +169,15 @@ const App = () => {
         />
         <ManageUsersModal />
       </div>
-            <ChemicalModal chemical={selectedChemical} show={showModal} handleClose={handleCloseModal}/>
-            <AddChemicalModal
-                show={showAddChemicalModal}
-                handleClose={handleCloseAddChemicalModal}
-            />
+      <ChemicalModal
+        chemical={selectedChemical}
+        show={showModal}
+        handleClose={handleCloseModal}
+      />
+      <AddChemicalModal
+        show={showAddChemicalModal}
+        handleClose={handleCloseAddChemicalModal}
+      />
     </div>
   );
 };
