+++ frontend/src/Navbar.js
@@ -1,0 +1,75 @@
+import { useState, useEffect } from "react";
+export const Navbar = ({ handleShowAddChemicalModal }) => {
+  const [user, setUser] = useState({});
+
+  useEffect(() => {
+    fetch("/api/user")
+      .then((response) => response.json())
+      .then((data) => setUser(data))
+      .catch((error) => console.error(error));
+  }, []);
+  return (
+    <nav className="navbar navbar-expand-lg bg-light">
+      <div className="container-fluid">
+        <a className="navbar-brand me-auto" href="/">
+          Chemical Inventory
+        </a>
+        <div className="collapse navbar-collapse" id="navbarSupportedContent">
+          <ul className="navbar-nav me-auto mb-2 mb-lg-0">
+            <li className="nav-item">
+              <a
+                className="nav-link active"
+                href="#"
+                onClick={(e) => {
+                  e.preventDefault();
+                  handleShowAddChemicalModal();
+                }}
+              >
+                Add Chemical
+              </a>
+            </li>
+          </ul>
+          <ul className="navbar-nav mb-2 mb-lg-0">
+            <li className="nav-item dropdown">
+              <a
+                className="nav-link dropdown-toggle"
+                href="#"
+                id="navbarDropdown"
+                role="button"
+                data-bs-toggle="dropdown"
+                aria-expanded="false"
+              >
+                Hi {user.name}
+              </a>
+              <ul
+                className="dropdown-menu dropdown-menu-end"
+                aria-labelledby="navbarDropdown"
+              >
+                <li className="dropdown-item">You have {user.access} access</li>
+                {user.access === "admin" ? (
+                  <li>
+                    <a
+                      className="dropdown-item"
+                      data-bs-toggle="modal"
+                      data-bs-target="#manageUsersModal"
+                    >
+                      Manage access
+                    </a>
+                  </li>
+                ) : null}
+                <li>
+                  <hr className="dropdown-divider" />
+                </li>
+                <li>
+                  <a className="dropdown-item" href="#">
+                    Logout
+                  </a>
+                </li>
+              </ul>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </nav>
+  );
+};
